requirements.txt
anthropic>=0.18.0
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
tenacity>=8.2.0
python-dotenv>=1.0.0
pytz>=2024.1
